Refactor: Move ROS2 publishers and subscribers to integrations module
